requirements.txt
chromadb 
langchain
tiktoken
chainlit
openai
python-dotenv